# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BetterPlay API",
        "description": "Aquí podrás ver todas las funciones de la API detalladamente",
        "version": "1.0.0"
    },
    "paths": {
        "/api/events/create": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Crea un Evento",
                "description": "Recibe home_id, away_id, home_odd, away_odd, tie_odd, date, finalDate y crea evento",
                "operationId": "f0e8b6a154986ce59fc18586b3f0f899",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "home_id": {
                                        "type": "integer"
                                    },
                                    "away_id": {
                                        "type": "integer"
                                    },
                                    "home_odd": {
                                        "type": "integer"
                                    },
                                    "away_odd": {
                                        "type": "integer"
                                    },
                                    "tie_odd": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "integer"
                                    },
                                    "finalDate": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento Creado Correctamente"
                    }
                }
            }
        },
...
        "/api/events/getEventById": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "Obtiene y devuelve un evento concreto",
                "description": "Recibe la id de un evento, lo busca y lo devuelve",
                "operationId": "66794e7eeb5abf9d9b42c3c4bb21edbd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento Encontrado Correctamente"
                    }
                }
            }
        },
        "/api/events/participateInBet": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "El Usuario participa en una apuesta",
                "description": "Recibe la id de un evento, la cantidad de monedas apostada y el equipo elegido y realiza la apuesta",
                "operationId": "e1ed8f3120132856235a2efdb0c2d97c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "eventId": {
                                        "type": "integer"
                                    },
                                    "coins": {
                                        "type": "integer"
                                    },
                                    "team_selected": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento Encontrado Correctamente"
                    }
                }
            }
        },
        "/api/events/finishEvent": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "Finalización de un evento",
                "description": "Recibe un array de json que contienen la id del evento, el resultado local y el visitante",
                "operationId": "34cc79faef19a8d073f2a86162cfc8a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "eventId": {
                                        "type": "integer"
                                    },
                                    "home_result": {
                                        "type": "integer"
                                    },
                                    "away_result": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento Encontrado Correctamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}